move process check to separate os file
